Conflicts:
	src/main/java/com/gbax/TopicsTestTask/dao/MessageDao.java
	src/main/java/com/gbax/TopicsTestTask/dao/TopicDao.java

+++ src/main/java/com/gbax/TopicsTestTask/controller/MessageController.java
@@ -1,12 +1,16 @@
 package com.gbax.TopicsTestTask.controller;
 
 import com.gbax.TopicsTestTask.dao.entity.Message;
+import com.gbax.TopicsTestTask.enums.Errors;
 import com.gbax.TopicsTestTask.service.MessageService;
 import com.gbax.TopicsTestTask.system.exception.EntityNotFoundException;
+import com.gbax.TopicsTestTask.system.exception.NotAuthorizedException;
+import com.gbax.TopicsTestTask.system.security.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
@@ -23,10 +27,23 @@ public class MessageController {
     @Autowired
     HttpServletRequest request;
 
-    @ExceptionHandler(Throwable.class)
+    @Autowired
+    SecurityService securityService;
+
+    @ExceptionHandler(EntityNotFoundException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
     @ResponseBody
     public void handleEntityNotFoundException(final EntityNotFoundException e, final HttpServletRequest request,
+                                   Writer writer) throws IOException {
+        writer.write(String.format(
+                "{\"error\":{\"java.class\":\"%s\", \"error\":\"%s\"}}",
+                e.getClass(), e.getError().getId()));
+    }
+
+    @ExceptionHandler(NotAuthorizedException.class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    @ResponseBody
+    public void handleNotAuthorizedException(final NotAuthorizedException e, final HttpServletRequest request,
                                    Writer writer) throws IOException {
         writer.write(String.format(
                 "{\"error\":{\"java.class\":\"%s\", \"error\":\"%s\"}}",
@@ -48,13 +65,19 @@ public class MessageController {
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
     public Message create(@PathVariable("topicId") Integer topicId,
-                          @RequestBody Message message) throws EntityNotFoundException {
+                          @RequestBody Message message) throws EntityNotFoundException, NotAuthorizedException {
+        if (securityService.getSecurityPrincipal() == null) {
+            throw new NotAuthorizedException(Errors.NOT_AUTHORIZED);
+        }
         return messageService.addMessageToTopic(topicId, message);
     }
 
     @RequestMapping(value = "/{topicId}/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.OK)
-    public void remove(@PathVariable("topicId") Integer topicId, @PathVariable("id") Integer id) throws InterruptedException, EntityNotFoundException {
+    public void remove(@PathVariable("topicId") Integer topicId, @PathVariable("id") Integer id) throws InterruptedException, EntityNotFoundException, NotAuthorizedException {
+        if (securityService.getSecurityPrincipal() == null) {
+            throw new NotAuthorizedException(Errors.NOT_AUTHORIZED);
+        }
         messageService.remove(id);
     }
 }
